refactor: ScheduleSection 훅으로 빼내어 로직 분리, 스케줄 추가할때 정렬
